+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Fixed
+- Fixed a permissions issue with element actions.
 - Fixed migration bug when updating from Craft 2 to Craft 3.7. (thanks @engram-design)
 
 ## 3.2.1 - 2022-07-18